Add minor features
Enable to test in offline
Link header to home
Fix styles
