+++ .env.example
@@ -1,1 +1,3 @@
 REACT_APP_BACK_URL=
+REACT_APP_USERNAME=
+REACT_APP_PASSWORD=
