+++ src/components/Sub/css/Feed.css
@@ -7,7 +7,7 @@
   border-radius: 20px
 }
 
-@media screen and (max-width: 1350px) {
+@media screen and (max-width: 1400px) {
   .sub-content {
     margin: auto;
   }
