+++ src/pages/css/main.css
@@ -8,7 +8,7 @@
   vertical-align: top;
 }
 
-@media screen and (max-width: 1350px) {
+@media screen and (max-width: 1400px) {
   .main-block {
     width: 100%;
   }
