+++ src/tools/auth.ts
@@ -1,6 +1,8 @@
 import axios from "axios";
 import { APIBase } from './api';
 
+axios.defaults.withCredentials = true; // For cookies
+
 interface UserCredential {
   username: string,
   password: string
@@ -10,6 +12,21 @@ const auth = async () => {
   interface IAPIResponse { isAuth: boolean, username: string };
   const { data } = await axios.get<IAPIResponse>(APIBase + "/auth");
   return data;
+}
+
+const findUsername = async ( username: string ) => {
+  try {
+    interface IAPIResponse { result: boolean };
+
+    const { data } = await axios.post<IAPIResponse>(APIBase + "/auth/find-user", { username });
+    if(data.result) {
+      return 1; // Found
+    } else {
+      return 0; // Not found
+    }
+  } catch (e) {
+    return -1; // Error
+  }
 }
 
 const signinHandler = async (credential: UserCredential, next: any) => {
@@ -46,4 +63,4 @@ const signupHandler = async (credential: UserCredential, next: any) => {
   }
 }
 
-export { auth, signinHandler, signupHandler };
+export { auth, findUsername, signinHandler, signupHandler };
